+++ src/components/UserPage/PortfolioList.jsx
@@ -4,32 +4,36 @@ import { Link } from "react-router-dom";
 import DashBoard from "./DashBoard";
 import { Button, Col, Container, Row } from "react-bootstrap";
 import { LinkContainer } from "react-router-bootstrap";
-import moment from 'moment';
+import { FaEdit, FaTrash, FaRupeeSign} from "react-icons/fa";
 const Portfolio = (props) => (
   <tr>
     <td>{props.portfolio.code}</td>
-    <td>{props.portfolio.purchase_price}</td>
+    <td><FaRupeeSign/>{props.portfolio.purchase_price}</td>
     <td>{props.portfolio.purchase_quantity}</td>
-    <td>{props.portfolio.sale_price}</td>
+    <td>{(props.portfolio.purchase_date).split("T")[0]}</td>
+    <td><FaRupeeSign/>{props.portfolio.purchase_quantity * props.portfolio.purchase_price}</td>
+    <td><FaRupeeSign/>{props.portfolio.sale_price}</td>
     <td>{props.portfolio.sale_quantity}</td>
-    <td>{moment(props.portfolio.purchase_date).format('DD-MMM-YYYY')}</td>
-   
-    <td>{isNaN(props.portfolio.sale_date) ? moment(props.portfolio.sale_date).format('DD-MMM-YYYY'): ''}</td>
+
+    <td>{isNaN(props.portfolio.sale_date) ? (props.portfolio.sale_date).split("T")[0]: ''}</td>
     <td>
       <Link
         className="btn btn-link"
         to={`/portfolio_stock/edit/${props.portfolio._id}`}
       >
-        Edit
+        <FaEdit />
       </Link>{" "}
       |
       <button
         className="btn btn-link"
         onClick={() => {
-          props.deletePortfolio(props.portfolio._id);
+          const confirmBox = window.confirm("Do you really want to delete ");
+          if (confirmBox === true) {
+            props.deletePortfolio(props.portfolio._id);
+          }
         }}
       >
-        Delete
+        <FaTrash color="red" />
       </button>
     </td>
   </tr>
@@ -87,12 +91,7 @@ export default function PortfolioList() {
       <DashBoard />
       <Container>
         <Row>
-          <Col
-            lg={16}
-            md={16}
-            sm={12}
-            className="p-5 m-auto shadow-sm rounded-lg .ml-3"
-          >
+          <Col xs={12} className="m-auto shadow-sm rounded-lg">
             <h3>Portfolio List</h3>
             <LinkContainer to="/addportfolio">
               <Button> Add Stock Portfolio </Button>
@@ -102,11 +101,12 @@ export default function PortfolioList() {
               <thead>
                 <tr>
                   <th>Code</th>
-                  <th>Buy Price</th>
-                  <th>Buy Quantity</th>
+                  <th>Purchase Price</th>
+                  <th>Purchase Quantity</th>
+                  <th>Purchase Date</th>
+                  <th>Purchase Value</th>
                   <th>Sell Price</th>
                   <th>Sell Quantity</th>
-                  <th>Buy Date</th>
                   <th>Sell Date</th>
                   <th>Action</th>
                 </tr>
